+++ SF3000/includes/saudiosubscriber.h
@@ -9,15 +9,15 @@
 
 typedef struct SAudioContext {
   Item           creatorTask;   /* who to signal when we're done initializing */
-  ulong          creatorSignal; /* signal to send for synchronous completion */
+  uint32_t       creatorSignal; /* signal to send for synchronous completion */
   long           creatorStatus; /* result code for creator */
   Item           threadItem;    /* subscriber thread item */
   void*          threadStackBlock; /* pointer to thread's stack memory block */
   Item           requestPort;   /* message port item for subscriber requests */
-  ulong          requestPortSignal; /* signal to detect request port messages */
+  uint32_t       requestPortSignal; /* signal to detect request port messages */
   DSStreamCBPtr  streamCBPtr;   /* stream this subscriber belongs to */
   TemplateRecPtr templateArray; /* ptr to array of template records */
-  ulong          allBufferSignals; /* or's signals for all instruments on all channels */
+  uint32_t       allBufferSignals; /* or's signals for all instruments on all channels */
   Item           outputTemplateItem; /* template item for making channel's output instrument */
   Item           envelopeTemplateItem; /* template item for making channel's envelope instrument(s) */
   Item           decodeADPCMIns; /* Support for playing ADPCM data */
